Move the bridge resolver tests onto the backend-neutral resolver
The ASM module carried eleven bridge resolution tests against its own
`BridgeResolver`, which is now only a facade over the backend-neutral
resolver in the core module. None of them overlapped with the four tests
the core module already had, so the scenarios only ever ran against one
backend even though the logic they cover is shared by all of them.

Move all eleven into the core module, restating the descriptor helper on
`core.TypeUtils` instead of ASM's `Type`, so the whole set now runs where
the implementation lives. No test is dropped: the two modules ran 195
tests before and run 195 after.
